requirements.txt
transformers>=4.45.0
torch>=2.0.0
pillow>=10.0.0
safetensors>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.43.0
sentencepiece>=0.1.99
protobuf>=3.20.0
einops>=0.6.0
